+++ Content.Server/Nuke/NukeCodePaperSystem.cs
@@ -16,7 +16,7 @@ namespace Content.Server.Nuke
         [Dependency] private ChatSystem _chatSystem = default!;
         [Dependency] private StationSystem _station = default!;
         [Dependency] private PaperSystem _paper = default!;
-        [Dependency] private FaxSystem _faxSystem = default!;
+        [Dependency] private ServerFaxSystem _faxSystem = default!;
 
         public override void Initialize()
         {
