+++ Content.Shared/Fax/FaxSystem.cs
@@ -1,102 +1,244 @@
-using Content.Server.Administration;
-using Content.Server.Administration.Managers;
-using Content.Server.Chat.Managers;
-using Content.Server.DeviceNetwork.Systems;
-using Content.Server.Popups;
-using Content.Server.Power.Components;
-using Content.Server.Tools;
 using Content.Shared.Administration.Logs;
 using Content.Shared.Containers.ItemSlots;
+using Content.Shared.Damage.Systems;
 using Content.Shared.Database;
 using Content.Shared.DeviceNetwork.Events;
+using Content.Shared.DeviceNetwork.Systems;
 using Content.Shared.Emag.Systems;
-using Content.Shared.Fax;
+using Content.Shared.Popups;
 using Content.Shared.Fax.Components;
-using Content.Shared.Fax.Systems;
-using Content.Shared.GameTicking;
-using Content.Shared.Interaction;
 using Content.Shared.Labels.Components;
 using Content.Shared.Labels.EntitySystems;
 using Content.Shared.Mobs.Components;
 using Content.Shared.NameModifier.Components;
 using Content.Shared.Paper;
 using Content.Shared.Power;
-using Content.Shared.Tools;
 using Content.Shared.UserInterface;
-using Robust.Server.GameObjects;
-using Robust.Shared.Audio;
 using Robust.Shared.Audio.Systems;
 using Robust.Shared.Containers;
-using Robust.Shared.Player;
-using Robust.Shared.Prototypes;
+using Robust.Shared.Serialization;
 
-namespace Content.Server.Fax;
-
-public sealed partial class FaxSystem : EntitySystem
+namespace Content.Shared.Fax;
+/// <summary>
+/// System for handling execution of a mob within fax when copy or send attempt is made.
+/// </summary>
+public abstract partial class FaxSystem : EntitySystem
 {
-    [Dependency] private IChatManager _chat = default!;
-    [Dependency] private IAdminManager _adminManager = default!;
+    [Dependency] protected ISharedAdminLogManager AdminLogger = default!;
+    [Dependency] private DamageableSystem _damageable = default!;
+    [Dependency] protected EmagSystem Emag = default!;
     [Dependency] private ItemSlotsSystem _itemSlotsSystem = default!;
-    [Dependency] private SharedAppearanceSystem _appearanceSystem = default!;
-    [Dependency] private SharedGameTicker _gameTicker = default!;
-    [Dependency] private PopupSystem _popupSystem = default!;
-    [Dependency] private DeviceNetworkSystem _deviceNetworkSystem = default!;
-    [Dependency] private PaperSystem _paperSystem = default!;
     [Dependency] private LabelSystem _labelSystem = default!;
-    [Dependency] private SharedAudioSystem _audioSystem = default!;
-    [Dependency] private ToolSystem _toolSystem = default!;
-    [Dependency] private QuickDialogSystem _quickDialog = default!;
-    [Dependency] private UserInterfaceSystem _userInterface = default!;
-    [Dependency] private ISharedAdminLogManager _adminLogger = default!;
     [Dependency] private MetaDataSystem _metaData = default!;
-    [Dependency] private FaxecuteSystem _faxecute = default!;
-    [Dependency] private EmagSystem _emag = default!;
-
-    private static readonly ProtoId<ToolQualityPrototype> ScrewingQuality = "Screwing";
+    [Dependency] private PaperSystem _paperSystem = default!;
+    [Dependency] private SharedAppearanceSystem _appearanceSystem = default!;
+    [Dependency] protected SharedAudioSystem AudioSystem = default!;
+    [Dependency] private SharedDeviceNetworkSystem _deviceNetworkSystem = default!;
+    [Dependency] protected SharedPopupSystem PopupSystem = default!;
+    [Dependency] private SharedUserInterfaceSystem _userInterface = default!;
 
     private const string PaperSlotId = "Paper";
 
-    public override void Initialize()
+    [SubscribeLocalEvent]
+    private void OnComponentInit(Entity<FaxMachineComponent> entity, ref ComponentInit args)
     {
-        base.Initialize();
-
-        // Hooks
-        SubscribeLocalEvent<FaxMachineComponent, ComponentInit>(OnComponentInit);
-        SubscribeLocalEvent<FaxMachineComponent, MapInitEvent>(OnMapInit);
-        SubscribeLocalEvent<FaxMachineComponent, ComponentRemove>(OnComponentRemove);
-
-        SubscribeLocalEvent<FaxMachineComponent, EntInsertedIntoContainerMessage>(OnItemSlotChanged);
-        SubscribeLocalEvent<FaxMachineComponent, EntRemovedFromContainerMessage>(OnItemSlotChanged);
-        SubscribeLocalEvent<FaxMachineComponent, PowerChangedEvent>(OnPowerChanged);
-
-        // Interaction
-        SubscribeLocalEvent<FaxMachineComponent, InteractUsingEvent>(OnInteractUsing);
-        SubscribeLocalEvent<FaxMachineComponent, GotEmaggedEvent>(OnEmagged);
-
-        // UI
-        SubscribeLocalEvent<FaxMachineComponent, AfterActivatableUIOpenEvent>(OnToggleInterface);
-        SubscribeLocalEvent<FaxMachineComponent, FaxFileMessage>(OnFileButtonPressed);
-        SubscribeLocalEvent<FaxMachineComponent, FaxCopyMessage>(OnCopyButtonPressed);
-        SubscribeLocalEvent<FaxMachineComponent, FaxSendMessage>(OnSendButtonPressed);
-        SubscribeLocalEvent<FaxMachineComponent, FaxRefreshMessage>(OnRefreshButtonPressed);
-        SubscribeLocalEvent<FaxMachineComponent, FaxDestinationMessage>(OnDestinationSelected);
+        _itemSlotsSystem.AddItemSlot(entity.Owner, PaperSlotId, entity.Comp.PaperSlot);
+        UpdateAppearance(entity);
     }
 
+    [SubscribeLocalEvent]
+    private void OnComponentRemove(Entity<FaxMachineComponent> entity, ref ComponentRemove args)
+    {
+        _itemSlotsSystem.RemoveItemSlot(entity.Owner, entity.Comp.PaperSlot);
+    }
+
+    [SubscribeLocalEvent]
+    private void OnMapInit(Entity<FaxMachineComponent> entity, ref MapInitEvent args)
+    {
+        // Load all faxes on map in cache each other to prevent taking same name by user created fax
+        Refresh(entity);
+    }
+
+    [SubscribeLocalEvent]
+    private void OnItemSlotChanged(Entity<FaxMachineComponent> entity, ref ContainerModifiedMessage args)
+    {
+        if (!entity.Comp.Initialized)
+            return;
+
+        if (args.Container.ID != entity.Comp.PaperSlot.ID)
+            return;
+
+        var isPaperInserted = entity.Comp.PaperSlot.Item.HasValue;
+        if (isPaperInserted)
+        {
+            entity.Comp.InsertingTimeRemaining = entity.Comp.InsertionTime;
+            _itemSlotsSystem.SetLock(entity.Owner, entity.Comp.PaperSlot, true);
+        }
+
+        UpdateUserInterface(entity);
+    }
+
+    [SubscribeLocalEvent]
+    private void OnPowerChanged(EntityUid uid, FaxMachineComponent component, ref PowerChangedEvent args)
+    {
+        var isInsertInterrupted = !args.Powered && component.InsertingTimeRemaining > 0;
+        if (isInsertInterrupted)
+        {
+            component.InsertingTimeRemaining = 0f; // Reset animation
+
+            // Drop from slot because animation did not play completely
+            _itemSlotsSystem.SetLock(uid, component.PaperSlot, false);
+            _itemSlotsSystem.TryEject(uid, component.PaperSlot, null, out var _, true);
+        }
+
+        var isPrintInterrupted = !args.Powered && component.PrintingTimeRemaining > 0;
+        if (isPrintInterrupted)
+        {
+            component.PrintingTimeRemaining = 0f; // Reset animation
+        }
+
+        if (isInsertInterrupted || isPrintInterrupted)
+            UpdateAppearance(uid, component);
+
+        _itemSlotsSystem.SetLock(uid, component.PaperSlot, !args.Powered); // Lock slot when power is off
+    }
+
+    [SubscribeLocalEvent]
+    private void OnEmagged(EntityUid uid, FaxMachineComponent component, ref GotEmaggedEvent args)
+    {
+        if (!Emag.CompareFlag(args.Type, EmagType.Interaction))
+            return;
+
+        if (Emag.CheckFlag(uid, EmagType.Interaction))
+            return;
+
+        args.Handled = true;
+    }
+
+    [SubscribeLocalEvent]
+    private void OnPingPayload(Entity<FaxMachineComponent> ent, ref DeviceNetworkPacketEvent<FaxPingPayload> args)
+    {
+        var isForSyndie = Emag.CheckFlag(ent.Owner, EmagType.Interaction) && args.Data.IsSyndicate;
+        if (!isForSyndie && !ent.Comp.ResponsePings)
+            return;
+
+        var pong = new FaxPongPayload
+        {
+            FaxName = ent.Comp.FaxName,
+        };
+
+        _deviceNetworkSystem.SendPacket(ent.Owner, args.SenderAddress, ref pong);
+    }
+
+    [SubscribeLocalEvent]
+    private void OnPongPayload(Entity<FaxMachineComponent> ent, ref DeviceNetworkPacketEvent<FaxPongPayload> args)
+    {
+        ent.Comp.KnownFaxes[args.SenderAddress] = args.Data.FaxName;
+        UpdateUserInterface(ent.Owner, ent.Comp);
+    }
+
+    [SubscribeLocalEvent]
+    private void OnPrintPayload(Entity<FaxMachineComponent> ent, ref DeviceNetworkPacketEvent<FaxPrintPayload> args)
+    {
+        Receive(ent, args.Data.Data, args.SenderAddress);
+    }
+
+    [SubscribeLocalEvent]
+    private void OnToggleInterface(EntityUid uid, FaxMachineComponent component, AfterActivatableUIOpenEvent args)
+    {
+        UpdateUserInterface(uid, component);
+    }
+
+    [SubscribeLocalEvent]
+    private void OnFileButtonPressed(EntityUid uid, FaxMachineComponent component, FaxFileMessage args)
+    {
+        args.Label = args.Label?[..Math.Min(args.Label.Length, FaxFileMessageValidation.MaxLabelSize)];
+        args.Content = args.Content[..Math.Min(args.Content.Length, FaxFileMessageValidation.MaxContentSize)];
+        PrintFile(uid, component, args);
+    }
+
+    [SubscribeLocalEvent]
+    private void OnCopyButtonPressed(EntityUid uid, FaxMachineComponent component, FaxCopyMessage args)
+    {
+        if (HasComp<MobStateComponent>(component.PaperSlot.Item))
+            Faxecute(uid, component); // when button pressed it will hurt the mob.
+        else
+            Copy(uid, component, args);
+    }
+
+    [SubscribeLocalEvent]
+    private void OnSendButtonPressed(EntityUid uid, FaxMachineComponent component, FaxSendMessage args)
+    {
+        if (HasComp<MobStateComponent>(component.PaperSlot.Item))
+            Faxecute(uid, component); // when button pressed it will hurt the mob.
+        else
+            Send(uid, component, args);
+    }
+
+    [SubscribeLocalEvent]
+    private void OnRefreshButtonPressed(EntityUid uid, FaxMachineComponent component, FaxRefreshMessage args)
+    {
+        Refresh(uid, component);
+    }
+
+    [SubscribeLocalEvent]
+    private void OnDestinationSelected(EntityUid uid, FaxMachineComponent component, FaxDestinationMessage args)
+    {
+        SetDestination(uid, args.Address, component);
+    }
+
+    // We can't predict power shutting off, so we just let the animation continue if power gets cut out.
+    // If that ever changes, have this animation pause cause it would be pretty funny.
     public override void Update(float frameTime)
     {
         base.Update(frameTime);
 
-        var query = EntityQueryEnumerator<FaxMachineComponent, ApcPowerReceiverComponent>();
-        while (query.MoveNext(out var uid, out var fax, out var receiver))
+        var query = EntityQueryEnumerator<FaxMachineComponent>();
+        while (query.MoveNext(out var uid, out var fax))
         {
-            if (!receiver.Powered)
-                continue;
-
             ProcessPrintingAnimation(uid, frameTime, fax);
             ProcessInsertingAnimation(uid, frameTime, fax);
             ProcessSendingTimeout(uid, frameTime, fax);
         }
     }
+
+    protected void UpdateAppearance(EntityUid uid, FaxMachineComponent? component = null)
+    {
+        if (!Resolve(uid, ref component))
+            return;
+
+        if (TryComp<FaxableObjectComponent>(component.PaperSlot.Item, out var faxable))
+            component.InsertingState = faxable.InsertingState;
+
+
+        if (component.InsertingTimeRemaining > 0)
+        {
+            _appearanceSystem.SetData(uid, FaxMachineVisuals.VisualState, FaxMachineVisualState.Inserting);
+            Dirty(uid, component);
+        }
+        else if (component.PrintingTimeRemaining > 0)
+            _appearanceSystem.SetData(uid, FaxMachineVisuals.VisualState, FaxMachineVisualState.Printing);
+        else
+            _appearanceSystem.SetData(uid, FaxMachineVisuals.VisualState, FaxMachineVisualState.Normal);
+    }
+
+    protected void UpdateUserInterface(EntityUid uid, FaxMachineComponent? component = null)
+    {
+        if (!Resolve(uid, ref component))
+            return;
+
+        var isPaperInserted = component.PaperSlot.Item != null;
+        var canSend = isPaperInserted &&
+                      component.DestinationFaxAddress != null &&
+                      component.SendTimeoutRemaining <= 0 &&
+                      component.InsertingTimeRemaining <= 0;
+        var canCopy = isPaperInserted &&
+                      component.SendTimeoutRemaining <= 0 &&
+                      component.InsertingTimeRemaining <= 0;
+        var state = new FaxUiState(component.FaxName, component.KnownFaxes, canSend, canCopy, isPaperInserted, component.DestinationFaxAddress);
+        _userInterface.SetUiState(uid, FaxUiKey.Key, state);
+    }
+
 
     private void ProcessPrintingAnimation(EntityUid uid, float frameTime, FaxMachineComponent comp)
     {
@@ -118,7 +260,7 @@ public sealed partial class FaxSystem : EntitySystem
         if (comp.PrintingQueue.Count > 0)
         {
             comp.PrintingTimeRemaining = comp.PrintingTime;
-            _audioSystem.PlayPvs(comp.PrintSound, uid);
+            AudioSystem.PlayPvs(comp.PrintSound, uid);
         }
     }
 
@@ -149,214 +291,18 @@ public sealed partial class FaxSystem : EntitySystem
         }
     }
 
-    private void OnComponentInit(EntityUid uid, FaxMachineComponent component, ComponentInit args)
+    protected void Faxecute(EntityUid uid, FaxMachineComponent component, DamageOnFaxecuteEvent? args = null)
     {
-        _itemSlotsSystem.AddItemSlot(uid, PaperSlotId, component.PaperSlot);
-        UpdateAppearance(uid, component);
-    }
-
-    private void OnComponentRemove(EntityUid uid, FaxMachineComponent component, ComponentRemove args)
-    {
-        _itemSlotsSystem.RemoveItemSlot(uid, component.PaperSlot);
-    }
-
-    private void OnMapInit(EntityUid uid, FaxMachineComponent component, MapInitEvent args)
-    {
-        // Load all faxes on map in cache each other to prevent taking same name by user created fax
-        Refresh(uid, component);
-    }
-
-    private void OnItemSlotChanged(EntityUid uid, FaxMachineComponent component, ContainerModifiedMessage args)
-    {
-        if (!component.Initialized)
+        var sendEntity = component.PaperSlot.Item;
+        if (sendEntity == null)
             return;
 
-        if (args.Container.ID != component.PaperSlot.ID)
+        if (!TryComp<FaxecuteComponent>(uid, out var faxecute))
             return;
 
-        var isPaperInserted = component.PaperSlot.Item.HasValue;
-        if (isPaperInserted)
-        {
-            component.InsertingTimeRemaining = component.InsertionTime;
-            _itemSlotsSystem.SetLock(uid, component.PaperSlot, true);
-        }
-
-        UpdateUserInterface(uid, component);
-    }
-
-    private void OnPowerChanged(EntityUid uid, FaxMachineComponent component, ref PowerChangedEvent args)
-    {
-        var isInsertInterrupted = !args.Powered && component.InsertingTimeRemaining > 0;
-        if (isInsertInterrupted)
-        {
-            component.InsertingTimeRemaining = 0f; // Reset animation
-
-            // Drop from slot because animation did not play completely
-            _itemSlotsSystem.SetLock(uid, component.PaperSlot, false);
-            _itemSlotsSystem.TryEject(uid, component.PaperSlot, null, out var _, true);
-        }
-
-        var isPrintInterrupted = !args.Powered && component.PrintingTimeRemaining > 0;
-        if (isPrintInterrupted)
-        {
-            component.PrintingTimeRemaining = 0f; // Reset animation
-        }
-
-        if (isInsertInterrupted || isPrintInterrupted)
-            UpdateAppearance(uid, component);
-
-        _itemSlotsSystem.SetLock(uid, component.PaperSlot, !args.Powered); // Lock slot when power is off
-    }
-
-    private void OnInteractUsing(EntityUid uid, FaxMachineComponent component, InteractUsingEvent args)
-    {
-        if (args.Handled ||
-            !TryComp<ActorComponent>(args.User, out var actor) ||
-            !_toolSystem.HasQuality(args.Used, ScrewingQuality)) // Screwing because Pulsing already used by device linking
-            return;
-
-        _quickDialog.OpenDialog(actor.PlayerSession,
-            Loc.GetString("fax-machine-dialog-rename"),
-            Loc.GetString("fax-machine-dialog-field-name"),
-            (string newName) =>
-        {
-            if (component.FaxName == newName)
-                return;
-
-            if (newName.Length > 20)
-            {
-                _popupSystem.PopupEntity(Loc.GetString("fax-machine-popup-name-long"), uid);
-                return;
-            }
-
-            if (component.KnownFaxes.ContainsValue(newName) && !_emag.CheckFlag(uid, EmagType.Interaction)) // Allow existing names if emagged for fun
-            {
-                _popupSystem.PopupEntity(Loc.GetString("fax-machine-popup-name-exist"), uid);
-                return;
-            }
-
-            _adminLogger.Add(LogType.Action,
-                LogImpact.Low,
-                $"{ToPrettyString(args.User):user} renamed {ToPrettyString(uid):tool} from \"{component.FaxName}\" to \"{newName}\"");
-            component.FaxName = newName;
-            _popupSystem.PopupEntity(Loc.GetString("fax-machine-popup-name-set"), uid);
-            UpdateUserInterface(uid, component);
-        });
-
-        args.Handled = true;
-    }
-
-    private void OnEmagged(EntityUid uid, FaxMachineComponent component, ref GotEmaggedEvent args)
-    {
-        if (!_emag.CompareFlag(args.Type, EmagType.Interaction))
-            return;
-
-        if (_emag.CheckFlag(uid, EmagType.Interaction))
-            return;
-
-        args.Handled = true;
-    }
-
-    [SubscribeLocalEvent]
-    private void OnPingPayload(Entity<FaxMachineComponent> ent, ref DeviceNetworkPacketEvent<FaxPingPayload> args)
-    {
-        var isForSyndie = _emag.CheckFlag(ent.Owner, EmagType.Interaction) && args.Data.IsSyndicate;
-        if (!isForSyndie && !ent.Comp.ResponsePings)
-            return;
-
-        var pong = new FaxPongPayload
-        {
-            FaxName = ent.Comp.FaxName,
-        };
-
-        _deviceNetworkSystem.SendPacket(ent.Owner, args.SenderAddress, ref pong);
-    }
-
-    [SubscribeLocalEvent]
-    private void OnPongPayload(Entity<FaxMachineComponent> ent, ref DeviceNetworkPacketEvent<FaxPongPayload> args)
-    {
-        ent.Comp.KnownFaxes[args.SenderAddress] = args.Data.FaxName;
-        UpdateUserInterface(ent.Owner, ent.Comp);
-    }
-
-    [SubscribeLocalEvent]
-    private void OnPrintPayload(Entity<FaxMachineComponent> ent, ref DeviceNetworkPacketEvent<FaxPrintPayload> args)
-    {
-        Receive(ent, args.Data.Data, args.SenderAddress);
-    }
-
-    private void OnToggleInterface(EntityUid uid, FaxMachineComponent component, AfterActivatableUIOpenEvent args)
-    {
-        UpdateUserInterface(uid, component);
-    }
-
-    private void OnFileButtonPressed(EntityUid uid, FaxMachineComponent component, FaxFileMessage args)
-    {
-        args.Label = args.Label?[..Math.Min(args.Label.Length, FaxFileMessageValidation.MaxLabelSize)];
-        args.Content = args.Content[..Math.Min(args.Content.Length, FaxFileMessageValidation.MaxContentSize)];
-        PrintFile(uid, component, args);
-    }
-
-    private void OnCopyButtonPressed(EntityUid uid, FaxMachineComponent component, FaxCopyMessage args)
-    {
-        if (HasComp<MobStateComponent>(component.PaperSlot.Item))
-            _faxecute.Faxecute(uid, component); // when button pressed it will hurt the mob.
-        else
-            Copy(uid, component, args);
-    }
-
-    private void OnSendButtonPressed(EntityUid uid, FaxMachineComponent component, FaxSendMessage args)
-    {
-        if (HasComp<MobStateComponent>(component.PaperSlot.Item))
-            _faxecute.Faxecute(uid, component); // when button pressed it will hurt the mob.
-        else
-            Send(uid, component, args);
-    }
-
-    private void OnRefreshButtonPressed(EntityUid uid, FaxMachineComponent component, FaxRefreshMessage args)
-    {
-        Refresh(uid, component);
-    }
-
-    private void OnDestinationSelected(EntityUid uid, FaxMachineComponent component, FaxDestinationMessage args)
-    {
-        SetDestination(uid, args.Address, component);
-    }
-
-    private void UpdateAppearance(EntityUid uid, FaxMachineComponent? component = null)
-    {
-        if (!Resolve(uid, ref component))
-            return;
-
-        if (TryComp<FaxableObjectComponent>(component.PaperSlot.Item, out var faxable))
-            component.InsertingState = faxable.InsertingState;
-
-
-        if (component.InsertingTimeRemaining > 0)
-        {
-            _appearanceSystem.SetData(uid, FaxMachineVisuals.VisualState, FaxMachineVisualState.Inserting);
-            Dirty(uid, component);
-        }
-        else if (component.PrintingTimeRemaining > 0)
-            _appearanceSystem.SetData(uid, FaxMachineVisuals.VisualState, FaxMachineVisualState.Printing);
-        else
-            _appearanceSystem.SetData(uid, FaxMachineVisuals.VisualState, FaxMachineVisualState.Normal);
-    }
-    private void UpdateUserInterface(EntityUid uid, FaxMachineComponent? component = null)
-    {
-        if (!Resolve(uid, ref component))
-            return;
-
-        var isPaperInserted = component.PaperSlot.Item != null;
-        var canSend = isPaperInserted &&
-                      component.DestinationFaxAddress != null &&
-                      component.SendTimeoutRemaining <= 0 &&
-                      component.InsertingTimeRemaining <= 0;
-        var canCopy = isPaperInserted &&
-                      component.SendTimeoutRemaining <= 0 &&
-                      component.InsertingTimeRemaining <= 0;
-        var state = new FaxUiState(component.FaxName, component.KnownFaxes, canSend, canCopy, isPaperInserted, component.DestinationFaxAddress);
-        _userInterface.SetUiState(uid, FaxUiKey.Key, state);
+        var damageSpec = faxecute.Damage;
+        _damageable.ChangeDamage(sendEntity.Value, damageSpec);
+        PopupSystem.PopupEntity(Loc.GetString("fax-machine-popup-error", ("target", uid)), uid, PopupType.LargeCaution);
     }
 
     /// <summary>
@@ -387,7 +333,7 @@ public sealed partial class FaxSystem : EntitySystem
 
         var payload = new FaxPingPayload
         {
-            IsSyndicate = _emag.CheckFlag(uid, EmagType.Interaction),
+            IsSyndicate = Emag.CheckFlag(uid, EmagType.Interaction),
         };
 
         _deviceNetworkSystem.SendPacket(uid, null, ref payload);
@@ -411,7 +357,7 @@ public sealed partial class FaxSystem : EntitySystem
 
         // Unfortunately, since a paper entity does not yet exist, we have to emulate what LabelSystem will do.
         var nameWithLabel = (args.Label is { } label) ? $"{name} ({label})" : name;
-        _adminLogger.Add(LogType.Action,
+        AdminLogger.Add(LogType.Action,
             LogImpact.Low,
             $"{ToPrettyString(args.Actor):actor} " +
             $"added print job to \"{component.FaxName}\" {ToPrettyString(uid):tool} " +
@@ -441,6 +387,7 @@ public sealed partial class FaxSystem : EntitySystem
         TryComp<LabelComponent>(sendEntity, out var labelComponent);
         TryComp<NameModifierComponent>(sendEntity, out var nameMod);
 
+        // Comment does not exist >:(
         // TODO: See comment in 'Send()' about not being able to copy whole entities
         var printout = new FaxPrintout(paper.Content,
                                        nameMod?.BaseName ?? metadata.EntityName,
@@ -458,7 +405,7 @@ public sealed partial class FaxSystem : EntitySystem
 
         UpdateUserInterface(uid, component);
 
-        _adminLogger.Add(LogType.Action,
+        AdminLogger.Add(LogType.Action,
             LogImpact.Low,
             $"{ToPrettyString(args.Actor):actor} " +
             $"added copy job to \"{component.FaxName}\" {ToPrettyString(uid):tool} " +
@@ -512,7 +459,7 @@ public sealed partial class FaxSystem : EntitySystem
 
         _deviceNetworkSystem.SendPacket(uid, component.DestinationFaxAddress, ref payload);
 
-        _adminLogger.Add(LogType.Action,
+        AdminLogger.Add(LogType.Action,
             LogImpact.Low,
             $"{ToPrettyString(args.Actor):actor} " +
             $"sent fax from \"{component.FaxName}\" {ToPrettyString(uid):tool} " +
@@ -521,7 +468,7 @@ public sealed partial class FaxSystem : EntitySystem
 
         component.SendTimeoutRemaining += component.SendTimeout;
 
-        _audioSystem.PlayPvs(component.SendSound, uid);
+        AudioSystem.PlayPvs(component.SendSound, uid);
 
         UpdateUserInterface(uid, component);
     }
@@ -537,7 +484,7 @@ public sealed partial class FaxSystem : EntitySystem
 
         var faxName = printout.SenderFaxName ?? Loc.GetString("fax-machine-popup-source-unknown");
 
-        _popupSystem.PopupEntity(Loc.GetString("fax-machine-popup-received", ("from", faxName)), uid);
+        PopupSystem.PopupEntity(Loc.GetString("fax-machine-popup-received", ("from", faxName)), uid);
         _appearanceSystem.SetData(uid, FaxMachineVisuals.VisualState, FaxMachineVisualState.Printing);
 
         if (component.NotifyAdmins)
@@ -579,12 +526,80 @@ public sealed partial class FaxSystem : EntitySystem
             _labelSystem.Label(printed, label);
         }
 
-        _adminLogger.Add(LogType.Action, LogImpact.Low, $"\"{component.FaxName}\" {ToPrettyString(uid):tool} printed {ToPrettyString(printed):subject}: {printout.Content}");
+        AdminLogger.Add(LogType.Action, LogImpact.Low, $"\"{component.FaxName}\" {ToPrettyString(uid):tool} printed {ToPrettyString(printed):subject}: {printout.Content}");
     }
 
-    private void NotifyAdmins(string faxName)
+    protected abstract void NotifyAdmins(string faxName);
+}
+
+[Serializable, NetSerializable]
+public enum FaxUiKey : byte
+{
+    Key
+}
+
+[Serializable, NetSerializable]
+public sealed class FaxUiState : BoundUserInterfaceState
+{
+    public string DeviceName { get; }
+    public Dictionary<string, string> AvailablePeers { get; }
+    public string? DestinationAddress { get; }
+    public bool IsPaperInserted { get; }
+    public bool CanSend { get; }
+    public bool CanCopy { get; }
+
+    public FaxUiState(string deviceName,
+        Dictionary<string, string> peers,
+        bool canSend,
+        bool canCopy,
+        bool isPaperInserted,
+        string? destAddress)
     {
-        _chat.SendAdminAnnouncement(Loc.GetString("fax-machine-chat-notify", ("fax", faxName)));
-        _audioSystem.PlayGlobal("/Audio/Machines/high_tech_confirm.ogg", Filter.Empty().AddPlayers(_adminManager.ActiveAdmins), false, AudioParams.Default.AddVolume(-8f));
+        DeviceName = deviceName;
+        AvailablePeers = peers;
+        IsPaperInserted = isPaperInserted;
+        CanSend = canSend;
+        CanCopy = canCopy;
+        DestinationAddress = destAddress;
+    }
+}
+
+[Serializable, NetSerializable]
+public sealed class FaxFileMessage : BoundUserInterfaceMessage
+{
+    public string? Label;
+    public string Content;
+    public bool OfficePaper;
+
+    public FaxFileMessage(string? label, string content, bool officePaper)
+    {
+        Label = label;
+        Content = content;
+        OfficePaper = officePaper;
+    }
+}
+
+public static class FaxFileMessageValidation
+{
+    public const int MaxLabelSize = 50; // parity with Content.Server.Labels.Components.HandLabelerComponent.MaxLabelChars
+    public const int MaxContentSize = 10000;
+}
+
+[Serializable, NetSerializable]
+public sealed class FaxCopyMessage : BoundUserInterfaceMessage;
+
+[Serializable, NetSerializable]
+public sealed class FaxSendMessage : BoundUserInterfaceMessage;
+
+[Serializable, NetSerializable]
+public sealed class FaxRefreshMessage : BoundUserInterfaceMessage;
+
+[Serializable, NetSerializable]
+public sealed class FaxDestinationMessage : BoundUserInterfaceMessage
+{
+    public string Address { get; }
+    public FaxDestinationMessage(string address)
+    {
+        Address = address;
     }
 }
